MMM Full app/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
databricks-sdk==0.28.0
databricks-sql-connector==3.3.0
pandas
openpyxl==3.1.2
reportlab==4.2.0
httpx==0.27.0
